// web/src/API.ts
export type CreateNoteInput = {
  id?: string | null,
  title?: string | null,
  content?: string | null,
};
// ...
export type UpdateNoteInput = {
  id: string,
  title?: string | null,
  content?: string | null,
};